[examples] Merge dev-node-state-style into examples/node-state-style
Now that 0.44.0 is released and ships DOMRenderExtension, replace the
mutation-listener based node-state-style example with the extension-based
implementation from dev-examples/dev-node-state-style. Drop the dev-only
copy. The dev-examples directory and integration test infrastructure
remain so future examples built on unreleased APIs can use them.
