--- dev-examples/dev-node-state-style/README.md
@@ -1,13 +1,0 @@
-# DEV Node State Style example
-
-Here we have an example that demonstrates how NodeState can be used with a
-DOMRenderExtension to override create and export behavior of any node.
-
-This example currently depends on unreleased features (v0.44+) and will not
-work outside of the monorepo.
-
-**Run it locally:** `pnpm i && pnpm run dev`
-
-[![Open in StackBlitz](https://developer.stackblitz.com/img/open_in_stackblitz.svg)](https://stackblitz.com/github/facebook/lexical/tree/main?file=dev-examples/dev-node-state-style/src/main.tsx&startCommand=pnpm%20run%20start:example%20dev-node-state-style)
-
-[![Open in StackBlitz-PR](https://developer.stackblitz.com/img/open_in_stackblitz.svg)](https://stackblitz.com/github/etrepum/lexical/tree/domextension-override-sort?file=dev-examples/dev-node-state-style/src/main.tsx&startCommand=pnpm%20run%20start:example%20dev-node-state-style)
