--- dev-examples/dev-node-state-style/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@lexical/dev-node-state-style-example",
-  "private": true,
-  "version": "0.43.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite -c vite.config.monorepo.ts",
-    "monorepo:dev": "vite -c vite.config.monorepo.ts",
-    "build": "tsc && vite build -c vite.config.monorepo.ts",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@ark-ui/react": "^5.36.0",
-    "@lexical/clipboard": "workspace:*",
-    "@lexical/extension": "workspace:*",
-    "@lexical/history": "workspace:*",
-    "@lexical/html": "workspace:*",
-    "@lexical/react": "workspace:*",
-    "@lexical/rich-text": "workspace:*",
-    "@lexical/selection": "workspace:*",
-    "@lexical/utils": "workspace:*",
-    "@zag-js/combobox": "^1.40.0",
-    "@shikijs/langs": "^4.0.2",
-    "@shikijs/themes": "^4.0.2",
-    "inline-style-parser": "^0.2.7",
-    "lexical": "workspace:*",
-    "lucide-react": "^1.8.0",
-    "prettier": "^3.8.1",
-    "react": "^19.2.5",
-    "react-dom": "^19.2.5",
-    "shiki": "^4.0.2"
-  },
-  "devDependencies": {
-    "@types/node": "^25.6.0",
-    "@types/react": "^19.2.14",
-    "@types/react-dom": "^19.2.3",
-    "@vitejs/plugin-react": "^6.0.1",
-    "cross-env": "^10.1.0",
-    "csstype": "^3.2.3",
-    "typescript": "^6.0.2",
-    "vite": "^8.0.8"
-  }
-}
