--- dev-examples/dev-node-state-style/src/icons/LICENSE.md
@@ -1,5 +1,0 @@
-Bootstrap Icons
-https://icons.getbootstrap.com
-
-Licensed under MIT license
-https://github.com/twbs/icons/blob/main/LICENSE.md
