+++ examples/node-state-style/package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "@ark-ui/react": "^5.6.0",
-    "@lexical/clipboard": "0.44.0",
+    "@lexical/extension": "0.44.0",
+    "@lexical/history": "0.44.0",
     "@lexical/html": "0.44.0",
     "@lexical/react": "0.44.0",
+    "@lexical/rich-text": "0.44.0",
     "@lexical/selection": "0.44.0",
     "@lexical/utils": "0.44.0",
     "@shikijs/langs": "^3.3.0",
