+++ examples/node-state-style/src/plugins/StyleViewPlugin.tsx
@@ -622,7 +622,7 @@ function isNotVendorProperty(item: string): boolean {
 }
 
 function useSuggestedStylesCombobox(props: CSSPropertyComboBoxProps) {
-  const [initialItems] = useState(getSuggestedStyleKeys);
+  const initialItems = useMemo(() => getSuggestedStyleKeys(), []);
   const [items, setItems] = useState(() =>
     initialItems.filter(isNotVendorProperty).join('\n'),
   );
